misc/requirements.txt
ansible~=2.5.15
PyYAML==3.13
